+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * skinupdate command to invalidate the database entry (Added new locale messages - you should reset your messages file)
 * Removed error message if message file already exists
 * Load embed message file per default
+* Fix NPE on skin reset
 
 ##### 1.6